Create tables for the entire initial data model
Replace the existing migration scripts with ones that create everything in the
initial version of the data model document.

Add documentation about how migrations work.

Add support for database-specific schema variants, though Postgres is hardwired
into the build process for now.

+++ src/main/kotlin/com/terraformation/seedbank/api/OrganizationController.kt
@@ -24,6 +24,6 @@ class OrganizationController(private val organizationFetcher: OrganizationFetche
   }
 }
 
-data class ListOrganizationsElement(val id: Int, val name: String)
+data class ListOrganizationsElement(val id: Long, val name: String)
 
 data class ListOrganizationsResponse(val organizations: List<ListOrganizationsElement>)
